Move component to common file
