+++ packages/example/src/pages/shallow-false.tsx
@@ -1,39 +1,13 @@
-import {
-  NextQueryParamProvider,
-  StringParam,
-  useQueryParam,
-  withDefault
-} from 'next-query-params';
-import {ChangeEvent} from 'react';
+import {NextQueryParamProvider} from 'next-query-params';
+import {ComponentProps} from 'react';
+import ShallowTest from '../components/ShallowTest';
 
-type ServerSideProps = {
-  query: Record<string, string>;
-};
-
-function ShallowFalsePage(props: ServerSideProps) {
-  const [name, setName] = useQueryParam('name', withDefault(StringParam, ''));
-
-  function onNameInputChange(event: ChangeEvent<HTMLInputElement>) {
-    setName(event.target.value);
-  }
-
-  return (
-    <div>
-      <label>
-        Server side query:
-        <span>{Object.values(props.query)}</span>
-      </label>
-      <label>
-        Name: <input onChange={onNameInputChange} value={name} /> ({name})
-      </label>
-    </div>
-  );
-}
+type ServerSideProps = ComponentProps<typeof ShallowTest>;
 
 export default function WithParamProvider(props: ServerSideProps) {
   return (
     <NextQueryParamProvider shallow={false}>
-      <ShallowFalsePage query={props.query} />
+      <ShallowTest query={props.query} />
     </NextQueryParamProvider>
   );
 }
